checkIsNumber checked

+++ ivanova.maria/B1/functions.cpp
@@ -18,7 +18,7 @@ void iva::fillRandom(double* array, int size)
 
 bool iva::checkIsNumber(const char* str)
 {
-  for (int i = 0; i < (strlen(str)); i++)
+  for (size_t i = 0; i < (strlen(str)); i++)
   {
     if ((str[i] == '\n') ||(!isdigit(str[i])))
     {
